Use CMake for cross plaform builds Make code platform independent Use OpenGL 4.1

+++ 3DEngine/src/texture/DepthTexture.h
@@ -12,7 +12,7 @@ public:
 
 	DepthTexture(int width, int height, Format format);
 
-	enum class CompareMode : GLint { RefToTexture =  GL_COMPARE_REF_TO_TEXTURE, Value = GL_NONE};
+	enum class CompareMode : unsigned int { RefToTexture =  static_cast<unsigned int>(GL_COMPARE_REF_TO_TEXTURE), Value = static_cast<unsigned int>(GL_NONE)};
 
 	/// Determines how the texture will be evaluated. Compare always returns 0 or 1 depending on the result of the depth test. Value returns the actual value
 	void SetCompareMode(const CompareMode& mode);
